% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\docType{package}
\name{osinfo-package}
\alias{osinfo}
\alias{osinfo-package}
\title{osinfo: Get the os info in a cross platform way}
\description{
Get the os info in a cross platform way.
}
\author{
\strong{Maintainer}: Devin Pastoor \email{devin.pastoor@gmail.com}

}
